#28 interface 통합 및 검색 조건 수정
