+++ src/component/page/Accounting/Disbursement/DisbursementDetailModal/DisbursementDetailModal.tsx
@@ -6,71 +6,20 @@ import { DisbursementDetailModalStyled } from "../styled";
 import { StyledTable, StyledTd, StyledTh } from "../../../../common/styled/StyledTable";
 import { Button } from "../../../../common/Button/Button";
 import NoImage from "../../../../../assets/noImage.jpg";
+import {
+    IDisbursementDetailModal,
+    IDisbursementDetailModalResponse,
+    IPostResponse,
+    IUserInfo,
+    IUserInfoResponse,
+    ICustList,
+    ICustListResponse,
+} from "../../../../../models/interface/Accounting/Disbursement";
 
 export interface IDisbursementDetailModalProps {
     resoNum?: number;
     onSuccess: () => void;
     setResoNum: (resoNum: undefined) => void;
-}
-
-export interface IDisbursementDetailModal {
-    resoNum: number;
-    applyId: string;
-    applyName: string;
-    applyDept: string;
-    custId: number;
-    custName: string;
-    groupCode: string;
-    grCodeNm: string;
-    acctCode: string;
-    acctCodeNm: string;
-    applyDate: string;
-    useDate: string;
-    disbContent: string;
-    amount: string;
-    apprYn: string;
-    apprDate: string;
-    eviMaterial: string;
-    phsycalPath: string;
-    logicalPath: string;
-    fileSize: number;
-    fileExt: string;
-}
-
-export interface IDisbursementDetailModalResponse {
-    disbDetail: IDisbursementDetailModal;
-}
-
-export interface IPostResponse {
-    result: string;
-}
-
-export interface IUserInfo {
-    loginId: string;
-    name: string;
-    deptName: string;
-}
-
-export interface IUserInfoResponse {
-    userInfo: IUserInfo;
-}
-
-export interface ICommonList {
-    dtl_cod: string;
-    dtl_cod_nm: string;
-}
-
-export interface ICommonListResponse {
-    commonList: ICommonList[];
-}
-
-export interface ICustList {
-    custId: number;
-    custName: string;
-}
-
-export interface ICustListResponse {
-    custList: ICustList[];
 }
 
 export const DisbursementDetailModal: FC<IDisbursementDetailModalProps> = ({ resoNum, onSuccess, setResoNum }) => {
